Backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
requests
python-multipart
sentence-transformers
faiss-cpu
PyMuPDF
numpy
openai
